aarch64: fix 32bpp in Picasso, reduced size of some vars

+++ src/od-pandora/sysconfig.h
@@ -528,7 +528,7 @@ typedef int32_t uae_atomic;
 #define M68K_SPEED_25MHZ_CYCLES 128
 
 typedef unsigned int WPARAM;
-typedef long LPARAM;
+typedef int LPARAM;
 typedef int SOCKET;
 #define INVALID_SOCKET -1
